requirements.txt
hydra-core==0.11.3
joblib==0.15.1
jupyter==1.0.0
jupyterlab==2.1.4
kaggle==1.5.6
matplotlib==3.2.1
mlflow==1.8.0
numpy==1.18.5
pandas==1.0.4
pylint==2.5.2
pytorch-lightning==0.7.6
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
tensorboard==2.2.2
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.1
